Renamed to backup_execute_command.rs.
